requirements.txt
ultralytics>=8.0.0
streamlit
Pillow
numpy
torch
